Move overkiz climate definitions
